requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
DateTime==4.3
Django==3.1.5
idna==2.8
newsapi-python==0.2.6
oauthlib==3.1.0
PySocks==1.7.1
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.25.11
zope.interface==5.2.0
fontawesome-free==5.15.3
